requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.3
django-leaflet==0.29.0
django-phonenumber-field==7.1.0
django-smtp-ssl==1.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
phonenumberslite==8.13.15
Pillow==10.0.0
plotly==5.15.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
